replace int's with enums.

+++ Src/MoneyFox.Shared.Tests/Helper/RecurringPaymentHelperTests.cs
@@ -38,7 +38,7 @@ namespace MoneyFox.Shared.Tests.Helper
             recurring.Amount.ShouldBe(payment.Amount);
             recurring.Category.Id.ShouldBe(payment.Category.Id);
             recurring.Type.ShouldBe(PaymentType.Income);
-            recurring.Recurrence.ShouldBe((int) PaymentRecurrence.Daily);
+            recurring.Recurrence.ShouldBe(PaymentRecurrence.Daily);
             recurring.Note.ShouldBe(payment.Note);
         }
 
@@ -61,7 +61,7 @@ namespace MoneyFox.Shared.Tests.Helper
             };
 
             var recurring = RecurringPaymentHelper.GetRecurringFromPayment(payment, true,
-                (int) PaymentRecurrence.Weekly, enddate);
+                PaymentRecurrence.Weekly, enddate);
 
             recurring.ChargedAccount.Id.ShouldBe(3);
             recurring.TargetAccount.Id.ShouldBe(8);
@@ -71,7 +71,7 @@ namespace MoneyFox.Shared.Tests.Helper
             recurring.Amount.ShouldBe(payment.Amount);
             recurring.Category.Id.ShouldBe(payment.Category.Id);
             recurring.Type.ShouldBe(PaymentType.Expense);
-            recurring.Recurrence.ShouldBe((int) PaymentRecurrence.Weekly);
+            recurring.Recurrence.ShouldBe(PaymentRecurrence.Weekly);
             recurring.Note.ShouldBe(payment.Note);
         }
 
@@ -94,7 +94,7 @@ namespace MoneyFox.Shared.Tests.Helper
             };
 
             var recurring = RecurringPaymentHelper.GetRecurringFromPayment(payment, true,
-                (int) PaymentRecurrence.Monthly, enddate);
+                PaymentRecurrence.Monthly, enddate);
 
             recurring.ChargedAccount.Id.ShouldBe(3);
             recurring.TargetAccount.Id.ShouldBe(8);
@@ -104,7 +104,7 @@ namespace MoneyFox.Shared.Tests.Helper
             recurring.Amount.ShouldBe(payment.Amount);
             recurring.Category.Id.ShouldBe(payment.Category.Id);
             recurring.Type.ShouldBe(PaymentType.Transfer);
-            recurring.Recurrence.ShouldBe((int) PaymentRecurrence.Monthly);
+            recurring.Recurrence.ShouldBe(PaymentRecurrence.Monthly);
             recurring.Note.ShouldBe(payment.Note);
         }
 
@@ -127,7 +127,7 @@ namespace MoneyFox.Shared.Tests.Helper
             };
 
             var recurring = RecurringPaymentHelper.GetRecurringFromPayment(payment, false,
-                (int) PaymentRecurrence.Yearly, enddate);
+                PaymentRecurrence.Yearly, enddate);
 
             recurring.ChargedAccount.Id.ShouldBe(3);
             recurring.TargetAccount.Id.ShouldBe(8);
@@ -137,7 +137,7 @@ namespace MoneyFox.Shared.Tests.Helper
             recurring.Amount.ShouldBe(payment.Amount);
             recurring.Category.Id.ShouldBe(payment.Category.Id);
             recurring.Type.ShouldBe(PaymentType.Expense);
-            recurring.Recurrence.ShouldBe((int) PaymentRecurrence.Yearly);
+            recurring.Recurrence.ShouldBe(PaymentRecurrence.Yearly);
             recurring.Note.ShouldBe(payment.Note);
         }
 
@@ -182,7 +182,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Monthly,
+                Recurrence = PaymentRecurrence.Monthly,
                 StartDate = new DateTime(2015, 08, dayOfMonth),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
@@ -225,7 +225,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Biweekly,
+                Recurrence = PaymentRecurrence.Biweekly,
                 StartDate = new DateTime(2015, 08, 25),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
@@ -245,7 +245,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Weekly,
+                Recurrence = PaymentRecurrence.Weekly,
                 StartDate = new DateTime(2015, 08, 25),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
@@ -265,7 +265,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Monthly,
+                Recurrence = PaymentRecurrence.Monthly,
                 StartDate = new DateTime(2015, 08, 25),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
@@ -285,7 +285,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Yearly,
+                Recurrence = PaymentRecurrence.Yearly,
                 StartDate = new DateTime(2015, 08, 25),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
@@ -305,7 +305,7 @@ namespace MoneyFox.Shared.Tests.Helper
             var recurringPayment = new RecurringPaymentViewModel
             {
                 Id = 4,
-                Recurrence = (int) PaymentRecurrence.Weekly,
+                Recurrence = PaymentRecurrence.Weekly,
                 StartDate = new DateTime(2015, 08, 25),
                 ChargedAccountId = 2,
                 ChargedAccount = account,
